Fixed warnings and do not compile what is not needed

+++ src/onda_jit_x86_64.h
@@ -6,8 +6,8 @@
 #include <stddef.h>
 #include <stdint.h>
 
-size_t onda_jit_x86_64(const onda_runtime_t* rt,
-                       uint8_t** out_machine_code,
-                       size_t* out_machine_code_size);
+int onda_jit_x86_64(const onda_runtime_t* rt,
+                    uint8_t** out_machine_code,
+                    size_t* out_machine_code_size);
 
 #endif // ONDA_JIT_X86_64_H
